requirements.txt
# === Core ===
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1

# === Twilio Integration ===
twilio==9.3.3

# === ElevenLabs SDK ===
elevenlabs==1.4.1

# === WebSockets & Async ===
websockets==12.0
aiohttp==3.10.5

# === Audio Processing ===
pyaudio==0.2.14
sounddevice==0.4.6

# === Utilities ===
requests==2.32.3
numpy==1.26.4

# === Development / Debugging ===
typing-extensions>=4.9.0
